casts 335 -> 254: the union device retires the {lo,hi} clock family
The "i64-halves-pun" class (33 casts) was parked as faithful because a plain
`i64` member raises the class alignment to 8 and moves every field after it.
`#pragma pack(4)` removes that objection outright: <Clock64.h> is an 8-byte
union at FOUR-byte alignment, so the whole value and its two dword halves both
get real names and nothing moves. Applied to CInGameIcon, CTimer, CGrunt,
CGruntVoice, CSBI_StatzTabGruntBar, CBattlezMapConfig and the CPlay clock band
(the last three as ANONYMOUS unions that keep the old half-names, so no other
lane's file needs an edit). i64-halves-pun: 33 -> 6, and all 6 left are in
files another lane owns.

The same device closes six more seams that were parked as byte- or API-forced:
  * CRezImage's BITMAPINFOHEADER + colour table ARE one GDI BITMAPINFO
  * BITMAPFILEHEADER's "BM" magic stamp (BmpFileHeaderStamp, 4 sites)
  * DDBLTFX's 25-dword zero loop (BltFxWords)
  * LOGPALETTE's variable-length colour table (LogPal256 gains its SDK arm;
    ImagePool's raw `char buf[1028]` + overlay is gone)
  * the 0x400-byte BMP colour table's byte/PALETTEENTRY views (Palette256)
  * the Blowfish 8-byte block's dword/byte/len-byte views (BlowfishBlock) -
    this one took BitStreamBlowfishDecode 99.98 -> 100.00 EXACT

Model fixes that were not casts at all:
  * g_ailDriver64 was declared HMDIDRIVER but is only ever the app HINSTANCE
    fed to FindResource/LoadResource/SizeofResource - renamed g_midiResModule
    and typed HINSTANCE (4 casts gone, no cast added)
  * CDDrawWorkerHost::WrapCoord takes LONG*, not i32*: retail hands it
    &rc.right/&rc.bottom of the RECT it then passes to DrawCount (0x15a712).
    The 10 other call sites declare their pair LONG to match.
  * RegistryHelper::GetValueString's in/out count IS the registry LPDWORD
  * CCheckpointDlg's message map was a placeholder `const i32 = 6205544`;
    it is a real AFX_MSGMAP {&CDialog::messageMap, _messageEntries} plus one
    ON_BN_CLICKED(0x53a) row - both now DATA-defined, GetMessageMap still 100%
  * CHashBase::FromLink and the Grunt/Battlez type-collection walks route
    through the container's existing one seam instead of repeating it
  * four id-keyed CMapPtrToPtr lookups now use MapLookupById

Blowfish also lost its last three casts: the S-box reload is `memcpy` (which
is what retail's single `mov ecx,0x400; rep movsd` IS), the cipher indexes
g_bfS[k][x], and InitializeBlowfish takes the key as `const char*`. That flips
which arm of the encipher/decipher mirror wall matches - measured four ways and
written into both @early-stop notes: the memcpy arm banks decipher's MAX
(99.50 -> 99.875) and takes InitializeBlowfish to EXACT, at the cost of
encipher's current % (its own MAX stays 100).

Cost: 18 functions moved down, 57 points total, 40 of them the blowfish mirror.
No new reloc defects (9, same set as main). data-tu-order/class-sizes/vtable
gates all clean.

+++ src/Gruntz/ActionOptionsMenuBar.cpp
@@ -214,8 +214,8 @@ i32 CActionOptionsMenuBar::Render() {
     if (!m_active) {
         return 1;
     }
-    i32 sx = m_screenX;
-    i32 sy = m_screenY;
+    LONG sx = m_screenX; // LONG: WrapCoord's pair type (it also fills RECT fields)
+    LONG sy = m_screenY;
     (g_gameReg->m_world->m_level->m_mainPlane)->WrapCoord(&sx, &sy);
 
     RECT r = g_gameReg->m_world->m_level->m_planeCtx;
