+++ src/Image/CImage.cpp
@@ -536,8 +536,8 @@ void CImage::RenderImage(CResolveNode* info, CDDrawSurfacePair* dst) {
     }
 
     // The plain-surface path (no flip, no owned sprite): compute + clip the rect, BltFast.
-    i32 x = m_originX - m_anchorX + info->m_plotDX + info->m_screenX;
-    i32 y = m_originY - m_anchorY + info->m_plotDY + info->m_screenY;
+    LONG x = m_originX - m_anchorX + info->m_plotDX + info->m_screenX;
+    LONG y = m_originY - m_anchorY + info->m_plotDY + info->m_screenY;
     if (info->m_flags & 0x40000) {
         info->m_level->m_mainPlane->WrapCoord(&x, &y);
     }
@@ -682,8 +682,8 @@ void CImage::RenderFrameClipped(
 // operand artifacts. Code bytes otherwise byte-exact (clip + struct-copy end).
 RVA(0x001538c0, 0x257)
 void CImage::BlitNorm(CResolveNode* info, CDDrawSurfacePair* dst) {
-    i32 x = info->m_screenX - m_originX - info->m_plotDX - m_anchorX;
-    i32 y = info->m_screenY - m_originY - info->m_plotDY - m_anchorY;
+    LONG x = info->m_screenX - m_originX - info->m_plotDX - m_anchorX;
+    LONG y = info->m_screenY - m_originY - info->m_plotDY - m_anchorY;
     if (info->m_flags & 0x40000) {
         info->m_level->m_mainPlane->WrapCoord(&x, &y);
     }
@@ -776,8 +776,8 @@ void CImage::BlitNorm(CResolveNode* info, CDDrawSurfacePair* dst) {
 // the WrapCoord/CopyRect reloc-name artifacts.
 RVA(0x00153b20, 0x270)
 void CImage::BlitFlipV(CResolveNode* info, CDDrawSurfacePair* dst) {
-    i32 x = info->m_screenX - info->m_plotDX - m_anchorX - m_originX;
-    i32 y = m_originY - m_anchorY + info->m_plotDY + info->m_screenY;
+    LONG x = info->m_screenX - info->m_plotDX - m_anchorX - m_originX;
+    LONG y = m_originY - m_anchorY + info->m_plotDY + info->m_screenY;
     if (info->m_flags & 0x40000) {
         info->m_level->m_mainPlane->WrapCoord(&x, &y);
     }
@@ -866,8 +866,8 @@ void CImage::BlitFlipV(CResolveNode* info, CDDrawSurfacePair* dst) {
 // WrapCoord/CopyRect reloc-name artifacts.
 RVA(0x00153d90, 0x259)
 void CImage::BlitFlipH(CResolveNode* info, CDDrawSurfacePair* dst) {
-    i32 x = info->m_plotDX - m_anchorX + m_originX + info->m_screenX;
-    i32 y = info->m_screenY - m_originY - m_anchorY - info->m_plotDY;
+    LONG x = info->m_plotDX - m_anchorX + m_originX + info->m_screenX;
+    LONG y = info->m_screenY - m_originY - m_anchorY - info->m_plotDY;
     if (info->m_flags & 0x40000) {
         info->m_level->m_mainPlane->WrapCoord(&x, &y);
     }
@@ -960,8 +960,8 @@ void CImage::BlitFlipH(CResolveNode* info, CDDrawSurfacePair* dst) {
 // reloc-name operand artifacts. Clip + inclusive-rect struct-copy end match.
 RVA(0x00153ff0, 0x280)
 void CImage::BlitShadeFlipHV(CResolveNode* info, CDDrawSurfacePair* dst) {
-    i32 x = info->m_screenX - m_anchorX + m_originX + info->m_plotDX;
-    i32 y = info->m_screenY - m_anchorY + m_originY + info->m_plotDY;
+    LONG x = info->m_screenX - m_anchorX + m_originX + info->m_plotDX;
+    LONG y = info->m_screenY - m_anchorY + m_originY + info->m_plotDY;
     if (info->m_flags & 0x40000) {
         info->m_level->m_mainPlane->WrapCoord(&x, &y);
     }
@@ -1059,8 +1059,8 @@ void CImage::BlitShadeFlipHV(CResolveNode* info, CDDrawSurfacePair* dst) {
 // same wall. Plus the WrapCoord ILT-thunk / CopyRect IAT-import reloc-name artifacts.
 RVA(0x00154270, 0x257)
 void CImage::BlitShadeNorm(CResolveNode* info, CDDrawSurfacePair* dst) {
-    i32 x = info->m_screenX - m_originX - m_anchorX - info->m_plotDX;
-    i32 y = info->m_screenY - m_originY - m_anchorY - info->m_plotDY;
+    LONG x = info->m_screenX - m_originX - m_anchorX - info->m_plotDX;
+    LONG y = info->m_screenY - m_originY - m_anchorY - info->m_plotDY;
     if (info->m_flags & 0x40000) {
         info->m_level->m_mainPlane->WrapCoord(&x, &y);
     }
@@ -1149,8 +1149,8 @@ void CImage::BlitShadeNorm(CResolveNode* info, CDDrawSurfacePair* dst) {
 // End-store struct-copy matches retail.
 RVA(0x001544d0, 0x275)
 void CImage::BlitShadeFlipV(CResolveNode* info, CDDrawSurfacePair* dst) {
-    i32 x = info->m_screenX - m_anchorX - info->m_plotDX - m_originX;
-    i32 y = m_originY + info->m_plotDY + info->m_screenY - m_anchorY;
+    LONG x = info->m_screenX - m_anchorX - info->m_plotDX - m_originX;
+    LONG y = m_originY + info->m_plotDY + info->m_screenY - m_anchorY;
     if (info->m_flags & 0x40000) {
         info->m_level->m_mainPlane->WrapCoord(&x, &y);
     }
@@ -1238,8 +1238,8 @@ void CImage::BlitShadeFlipV(CResolveNode* info, CDDrawSurfacePair* dst) {
 // operand artifacts. Clip + inclusive-rect struct-copy end match retail.
 RVA(0x00154750, 0x275)
 void CImage::BlitShadeFlipH(CResolveNode* info, CDDrawSurfacePair* dst) {
-    i32 x = info->m_plotDX + m_originX + info->m_screenX - m_anchorX;
-    i32 y = info->m_screenY - m_originY - info->m_plotDY - m_anchorY;
+    LONG x = info->m_plotDX + m_originX + info->m_screenX - m_anchorX;
+    LONG y = info->m_screenY - m_originY - info->m_plotDY - m_anchorY;
     if (info->m_flags & 0x40000) {
         info->m_level->m_mainPlane->WrapCoord(&x, &y);
     }
